Contain files within a subdirectory
This will keep everything in one place and allow simpler maintenance in the
future.

The idea is that users will simply fork the `dotparty` repo, then add their own
custom dotfiles to it. Keeping all the core stuff in one directory will allow
for seamless updates from the upstream repository, since there shouldn't be any
collisions with user files.

The chances that the user will want to have their own `party` directory are
slim, and they can link to one on their own if they wish.